feat(*): new reducer entity instance support
Allow the usage of non-registered entities which considered to be self-contained entities.

Entity type creation is also simplified by removing the need for a special instance to be created off of an entity type, a custom constructor name is now created based on the entity id (type:name).

BREAKING CHANGE: Entity type factory, custom entities, deprecated transform and source entity types. When creating entity types, the spec has to be passed as a property of the entity. `DataPoint.createEntity` API changed, it now expects the third argument to be `resolve:Function`.