feat(data-point): removes acc.resolve from middleware, new API introduced
to exit the middleware chain pass the new value as a second parameter

BREAKING CHANGE: API changed, middleware acc.resolve is no longer available it will be repurposed
later on. To achieve previous funcitonality you must pass a second parameter to the `next()` method.
See README for new API and examples on how to use it.

closes #206